chore(mocks/constants): shift constants used in stories to mocks/constants
